events: Extend the events_unwatch_fd() API with an event type
If a single file descriptor is being watch for multiple event types, it
will be present in multiple entries in the watch list. The
events_unwatch_fd() function will remove the first entry regardless of
its type, not allowing removal of a particular event type.

Fix this by extending the events_unwatch_fd() API to take an event type,
and remove the corresponding entry.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>