events: Remove select timeout
The reason why no timeout is needed is twofold:

When the stream is off extensive periods of time can pass without an
event occurring if the host doesn't send any UVC request. As that period
of time is unbound, we can't set any meaningful timeout.

When the stream is on buffers are supposed to flow constantly. If
buffers stop flowing that would indicate a bug somewhere in the stack,
which could be caught by a timeout, but that we wouldn't be able to
recover from anyway.

The timeout is thus not needed and harmful when the stream is off.
Remove it.

Signed-off-by: Paul Elder <paul.elder@ideasonboard.com>
Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>