uvc-gadget: Add struct uvc_stream
Since we are preparing to stream frames from a capture v4l2 device to
uvc-gadget, add a struct to represent this "frame-forwarding" stream,
such that it contains a representation of the uvc gadget device (struct
uvc_device) and other data relative to the stream, such as a pointer to
the events handler and test pattern generation data. A capture v4l2
device object will be added in a subsequent modification.

The stream is allocated dynamically, and the event handler is created
separately from the stream. This will allow replacing the event handling
mechanism with a different implementation without impacting the stream
implementation.

Change the argument to functions that need to be able to access stream
data from uvc_device to uvc_stream.

Signed-off-by: Paul Elder <paul.elder@ideasonboard.com>
Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>