configfs: Provide configfs support
Parse a configuration name from the commandline, and utilise it to
identify the configfs configuration path.

Only the short-name (i.e. "uvc.1") is necessary to provide if there is
no ambiguity regarding the gadget, otherwise the gadget path should be
included ("g1/functions/uvc.1"). If the parameter is not provided then
the first function is utilised.

Legacy g_webcam is still supported, and the parameter will define the
UDC to match against if provided.

Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>
Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>