v4l2: Extract video buffer structure to separate file
The v4l2_video_buffer structure describes a video buffer. It is used by
the v4l2_device class only, but isn't otherwise tied to V4L2. To prepare
for non-V4L2 video sources, extract it to a separate file and rename it
to video_buffer.

At the same time, add a new video_buffer_set structure to represent as
set of video buffers.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>