lib: Import glob.c from glibc 2.26
The Android libc implementation (bionic) doesn't provide the glob()
function. To enable compilation of the UVC gadget library for Android,
import the glibc glob() implementation.

The glob.c and glob.h files are imported unmodified to ease copyright
management, even if they don't compile as-is. Compilation will be fixed
separately.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>