configfs: parse mjpeg streaming format
The MJPEG descriptor does not have a guidFormat, and so configfs doesn't
have it either, but we currently error out if there is no guidFormat.

To fix this, extract the stream type part of the real path, and if it is
MJPEG then hardcode it, otherwise read the guid if it is uncompressed.
This does mean that in configfs the MJPEG format descriptor path must be
"mjpeg" for this to work, though this is already guranteed by configfs.

Signed-off-by: Paul Elder <paul.elder@ideasonboard.com>
Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>