lib: configfs: Use wildcard in path to gadget device
At some point the kernel has been updated to include a .id descriptor
in the gadget directory name - add that to uvc_find_video_device() so
that the video node is located properly.

Rather than hardcode a specific gadget (for instance with device name
gadget.0) use a wildcard. This has the advantage of backwards
compatability with older kernels that did not use the ID as part of the
device name, though still limits us to supporting scenarios where only a
single gadget is bound to the UDC.

Reviewed-by: Kieran Bingham <kieran.bingham@ideasonboard.com>
Signed-off-by: Daniel Scally <dan.scally@ideasonboard.com>