lib/jpg-source: Constrain fps
As written the fps output by the jpg-source is unconstrained; it will
simply fill buffers as quickly as the USB subsysem can handle them.
This is not particularly realistic and also means that the frame rates
configured (and thus expected) by the host are not being adhered to.
Use the new timer infrastructure to slow down the buffer fill operation
such that the frame rates seen match those configured by the host.

Reviewed-by: Kieran Bingham <kieran.bingham@ideasonboard.com>
Signed-off-by: Daniel Scally <dan.scally@ideasonboard.com>